[SERVICE-865] Test Updates
